Splitted components into their own files and added history tracking
