+++ tic-tac-toe/src/Square.js
@@ -1,0 +1,3 @@
+export function Square({value, onSquareClick}){
+    return <button className="square" onClick={onSquareClick}>{value}</button>;
+  }
